requirements.txt
certifi==2019.9.11
chardet==3.0.4
cmdy==0.1.9
colorama==0.4.1
completions==0.0.7
Django==2.2.5
django-embed-video==1.3
django-multiselectfield==0.1.8
django-taggit==1.1.0
idna==2.8
modkit==0.0.8
Pillow==6.1.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
pyparam==0.2.3
python-box87==3.9.10
python-simpleconf==0.1.10
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.3
wcwidth==0.1.7
